Implement the expensive tests for AbstractFeatureTest
The smaller versions of the 2D and 3D test images are created, and the
original ones are renamed with "_expensive" suffix. The tests in
GeomTest are modified such that each has different expected values for
expensive and inexpensive conditions. The expected values for the
smaller test images are calculated using the current implementations of
the geom package. Any future test class that uses the 2D or 3D test
image should do similarly.