Make op implementations extend proper abstract SpecialOp
Although many ops are modified, the changes are similar so I do not
split the commit. All of them include changing "extends AbstractOp" to
"extends AbstractNullary(Unary)FunctionOp", renaming variables, and
removing redundant methods/variables/imports. The logic of the classes
are not changed, and the way to retrieve these ops remains the same.
This is achieved by only making the required parameters as inputs.