Add real support for binary inplace ops
These ops compute a result from two arguments, storing the
result back into one of the two arguments.

The BinaryInplace1Op#mutate1 method stores it into the 1st one.
The BinaryInplaceOp#mutate2 method stores it into the 2nd one.

The BinaryInplaceOp extends BinaryInplace1Op, so implementers
can create either:

A) a binary inplace op which must mutate the first argument; or
B) a binary inplace op which can mutate either argument.

In both cases, when the op is used unspecially via the run() method,
the binary inplace op will mutate the first argument, not the second.

Neither kind of binary inplace op is capable of mutating both arguments
at once, and the two input arguments must be different because otherwise
there may be problems similar to computer ops when in == out.

This commit is dedicated to Leon Yang!