Rename UnaryInplaceOp accessor and mutator
* arg() -> in()
* setArg(A arg) -> setInput(A in)

The next step is for UnaryInplaceOp to extend UnaryOp, inheriting that
same API, eliminating the need for these extra methods altogether.

I used the name "arg" before to avoid potential confusion, since the
argument is both the input _and_ the output. But with the coming
refactoring and addition of BinaryInplaceOp, we want to reuse the
UnaryOp and BinaryOp interfaces. So the name will have to change.