Add a run(...) convenience layer
The computer ops have a compute method; the function ops have a compute
method too but slightly different; the inplace ops have a mutate method.

The run convenience layer unifies these various computation methods in a
(hopefully) intuitive way. Callers can now pass arguments to these run
methods in various ways, and the framework will try to do the most
correct and intuitive thing for the situation.

For example, if you call `run(input, null)` on a unary hybrid op, it
will call the method:

    UnaryFunctionOp#compute(input)

returning the result. But if you call `run(input, output)` with a
non-null `output` reference, it will instead call:

    UnaryComputerOp#compute(input, output)

and return that same `output` reference afterward.

Various combinations of computer, function and inplace are now handled
at each interface layer to maximize convenience for the caller.

Note that there are a few TEMP code blocks and hacks that will
disappear in subsequent commits, as some other related APIs are
extended and enhanced.