Move OpUtils#run into OpEnvironment
This helper method (as well as OpUtils#result) was only used
within the default interface method implementations of
OpEnvironment. The OpUtils class is rather a hodgepodge which is
best kept tidy (maybe even eliminated before the 1.0.0 release).
So it's cleaner to put this very special-purpose run method into
the OpEnvironment directly, and kill the unneeded result method.