WIP: Add a swanky fast OpIndex data structure
Like several other SJC-based indices, this one is built on the
SortedObjectIndex, a structure that keeps sorted lists by type.

In the case of ops, we categorize each OpInfo by its concrete op class,
meaning they all get loaded when the OpIndex is populated. But this is a
one-time cost, whereas op lookups by type become O(1) each.

In theory this is a big win, but in practice the performance boost might
be dwarfed by the cost of reflection later instantiating the candidates.
Further testing is needed. There are also a few failing unit tests...