Rename ThresholdUtils utility class to Thresholds
This also makes the class final and adds a private constructor,
to prevent misunderstanding of the class's intended usage.

This is consistent with the ImgLib2 way of defining utility classes.
Personally I think it reads more nicely and succinctly, without
sacrificing the ability to infer that it is a utility class.