Fix generics for create nativeType & imgFactory
When no Class<T> is given to create.nativeType, DoubleType is returned.
This scenario is now implemented as its own op, to avoid an
untenable T parameter in the CreateNamespace#nativeType() method.

The create.imgFactory op is trickier at the moment, because the
ImgFactory class takes a T parameter that is not anchored by any
constructor argument. The only place ImgFactory uses its T is as a
constraint to the type passed to its create method. So in actuality,
there is no direct harm per se in coercing the
CreateNamespace#imgFactory() and CreateNamespace#imgFactory(Dimensions)
methods to whatever T is specified by the caller; it merely limits the
flexibility of subsequent calls to ImgFactory#create(T).

See also: https://github.com/imglib/imglib2/issues/91