Distinguish compute threshold methods and their apply ops
For each threshold method, there is a "compute" op which outputs the
threshold value, and a "apply" op which applies the computed threshold
value to an II/RAI, and both have the same input signature. By using
priority, now the "compute" op is matched if no output object is
provided.