Fix generics of filter.gauss ops
We cannot have a separate V output type when the input type is T.
Not with the ops being hybrid. Because there is no way at runtime
of creating the RandomAccessibleInterval<V> output without having
an instance of type V.

If we want to support gauss with different input and output types,
it must be a ComputerOp, not a HybridCF. We can certainly do that
in a separate commit, but for the moment, constraining the output
type to match the input type is the least involved change.