DefaultSize: avoid using getRealDouble/setReal
Accumulating the count in a double is prone to rounding error. Let's
instead use the built-in functions of the output type to accumulate.

While this may be slower, the default implementation should always opt
for correctness and generality over performance. We can achieve better
performance by creating a second more specialized and optimized op.