WIP: MathNamespace: consolidate +-/* op signatures
[This is a WIP because it causes the completeness test to fail.]

This is by no means perfect: we still have six signatures for each
of the four operations (+-/*), due to the six (BDFILS) primitive types.

However, there is no need to keep both ArrayImg-specific and
PlanarImg-specific signatures separate. It is extremely rare that a
downstream caller would care what sort of Img is being used. One of
the primary goals of ops is to hide this complexity behind common
higher-level image interfaces.

So ultimately, we will probably want to move away from the Img-specific
signatures altogether, in favor of IterableInterval and/or
RandomAccessibleInterval. But that is a problem for another day.