Use ExtendedRAI wherever applicable
ExtendedRAI should be used if access of out of bound data is necessary.
The caller of an op that needs OOB data access should determine the type
of OOBF to use. Callee could use ExtendedRAI#getSource() to retrieve the
original RAI, and use another OOBF to create an ExtendedRAI if
necessary. With this new protocol, the RAIs#extends is no more needed.