Add test for ScaleImg
The test scales a small CellImg up and checks if the output is NOT a
CellImg (potentially an ArrayImg, which is the default type for small
images). A better test would be scaling a large ArrayImg up and check if
the output is a CellImg, but this will be too slow.