OpMatchingService: rejigger findMatches method
It is now called filterMatches, and instead of returning a List<Module>,
returns a List<OpCandidate> which is a filtered subset of the input
List<OpCandidate>. This is more flexible, since each OpCandidate does
contain a reference to its associated Module instance, but also the
other details of its candidacy.

And the old findMatches method still exists, deprecated, for backwards
compatibility.