Deprecate findModule in OpMatchingService
The usage of findModule is replaced by findMatch, which returns an
OpCandidate. A Module could be obtained by calling
findMatch(...).getModule().

Note that we do not need to deprecate

   findModule(OpEnvironment, List<OpRef<?>>)

because it was just added in e2a2004a4cd0119173f2db47df6e9cb650532220
and never present in a release.