Unit tests: use explicit op class for testing
Since the OpMethods in the namespaces no longer match the exact op class
as indicated in its annotation, testing should not rely on OpMethods,
except the method is called as a helper and is not intended to be
tested, i.e. using ops.stats().sum() to calculate the sum of a filtered
image, or when there is a dedicated test for checking if the OpMethods
are working correctly.