Improve Op matching mechanism
DefaultOpMatchingService now do a multi-pass matching instead of a
single pass.

First it search for Ops whose arguments match exactly with the
references.

If fails, it search for Ops whose arguments match by casting. Ops with
fewest total levels of casting will be returned.

If fails, it then search for Ops whose arguments match by lossless
conversion. This functionality is not yet implemented, because it needs
support from the conversion service to tell if a conversion is lossy.

If still no match is found, it does the regular search.