Move extension of integral images to thresholds
Per definition an integral image should have the same size as the input
image from which it is constructed. However, for the IntegralMean and
IntegralVariance ops to work properly the integral images has to be
extended with 0s in north and west and the boundary intensities in east
and south directions (2D case).