LocalThresholdIntegral: use OOBF when provided
This changes the behavior of the integral image-backed local thresholding
techniques to behave like their normal counterparts. That is, always use an
OOBF even if none is provided.

This might however skew the results of computations near the boarders. The
replaced computation took into account the overlap between the image and the
currently processed neighborhood. While this ignores a provided OOBF it computes
more correct values near the image boundary.