Make OpService optional for geom converters
This works around a problem where the creation of a limited context
which includes a ConvertService but not an OpService will spout
exceptions, since the two converters in question here are Converter
plugins which demand an OpService. Nicer is not to demand it, but
rather to request it (via required = false), and then fail
gracefully when no OpService is available.