Reduce size and complexity of PNG test images
The PNG files should be grayscale, not RGB. This is helpful both
in terms of storage space, as well as complexity when reading in
the data: they will now all be treated as single-channel images.

Somehow, these conversions affected the tests. I guess that certain
pixel values of the images were slightly altered. We are no longer
going through the weighted ARGB -> Float conversion specified by
the ARGBToFloatConverter; in theory, this shouldn't matter if R=G=B
in every case, but maybe there was rounding error sometimes.
So this commit also updates the expected values of those tests.