OpRef: relax the constraint on op type
While it intuitively makes sense to only allow op types that extend the
Op interface, there is no explicit need to constrain the primary type
in this way. And we will be switch away from Class soon toward Type,
so it is good to verify that everything is OK with arbitrary classes.