bugfix for reported bug on
http://forum.imagej.net/t/issue-with-polygons-in-ops/2227

The "inverted" variable was set to false, but in a LabelRegionCursor, only pixels with value=true exist. That's why never a polygon was created if a LabelRegion (aka RandomAccessibleInterval<BoolType>) was handed over as binary image.

In case of handing over a RandomAccessibleInterval<BitType>, a polygon was created, which corresponded to the outline of the image and no structure inside it.