Add interfaces for pixel-based learners
Both SupervisedLearners and UnsupervisedLearners are learning a model
from input data that can be applied to individual pixels. That is, a
model is a UnaryComputerOp that computes and output value for an input
pixel.

A simple example for an UnsupervisedLearner is thresholding. Given an
image (with help of the histogram), generate a UnaryComputerOp that
compares the pixel value to the computed threshold and sets the output
pixel "true" or "false" (output is of BooleanType).