OpUtils: filter injectable inputs
Service and Context fields get populated when Context.inject is called
on the object. We do not want to treat such parameters as real inputs.
Previously, they were not, but with a forthcoming change to SciJava
Common, they will be included in the list of input parameters. So we
take proactive steps here to ensure ops matching continues to work.