Refactor AbstractIntegralImg to avoid duplications
The computers for that are summing up values along a dimension are
decoupled from their surrounding classes. AbstractIntegralImg delegates
the creation of those classes to its subclasses for each axis
individually, i.e. getComputer() is now called with a dimension (axis).

This includes a fix for the computation of squared integral images:
squaring of input values only needs to be done in one dimension;
computations along other axes are just summations.