Rename create.object to create.instance
This change avoids a confusing name clash with java.lang.Object.

The name "create.instance" is slightly less general, but accurately
conveys the current op behavior, which is to instantiate an object from
a class. One can imagine something similar for Types in general, not
just Classes -- and perhaps even for other kinds of inputs.