Use the Ops type hierarchy for dynamic access
Now, you can have an instance of net.imagej.ops.Ops backed by a
particular OpEnvironment. It has an accessor for each namespace,
which are instances of the nested static namespace classes (e.g.,
Ops.Convert). These classes in turn have one method per op, each of
which accepts an Object... and returns an Object.

In this way, the dynamic op access can be separated from the
OpEnvironment itself.

The downside is that this is still statically generated. Much better
would be to perform runtime code generation, such that all available
ops are accessible via script languages.