FilterNamespace: use base op interfaces
The decision we made for namespaces was that the type-safe namespace
method signatures would invoke ops via each base op interface.
In this way, the built-in methods are not tied to specific
implementations, but rather use the full power of the matcher to
choose the best available op for the arguments in question.