Make ifft computer really a computer
IFFT computer tries to save memory by writing temporary IFFT
calculations to the input image. This means the computer changes the
input. We can avoid this by making a copy of the input and using that
for the temporary calculations.

This will make the op use more memory and perform slower. In the future
an op type which allows using the input img for temporary calculations
should be created.

Signed-off-by: Curtis Rueden <ctrueden@wisc.edu>