Enable the possibility of looser namespace typing
Sometimes, there are multiple ops with slightly different typing needs,
that the namespace checker is too dumb to understand right now.

It would be quite complicated to make it smart enough to understand
these cases. So instead, for now, we let the programmer chose to disable
the strict namespace type checking on a case-by-case basis.

See next commit for an example of why this is necessary.