Make the thinning ops faster
This was the result of a super quick performance analysis, addressing
only the very lowest hanging fruit in the most expedient way. There is
very likely still much room for improving performance of these ops.

But already, in our quick test, morphological thinning was 4x
faster, and the other three strategies were 2x faster. So \^_^/!

This commit is dedicated to Leo Epstein.