Make FFT memory in AbstractFFTFilterC optional
Previously FFT memmory buffers in the FFT Filter Computers (such as
RichardsonLucyC and ConvolveFFTC) were required.  This meant the
user had to create memory before callng them.  It also made it harder to
match GPU filters (where the memory would be created at a lower level,
by the GPU algorithm).  The FFT memory parameters are now optional.
If they are not passed in, the Computers will create memory for the
FFTs.