Update derivativeGauss tests
TooManyDimensions was deleted, since it was no longer necessary.
TooFewDimensions was refactored into a test where the number of sigmas
passed differs from the number of image dimensions (this test is
expected to fail). The regression test was updated with a sigma array
instead of a single sigma for all dimensions.