Reduce array lookups
On every Octant filling the neighbourhood array is checked for
emptiness. A single foreground pixel is sufficient to trigger a not
empty state, so the search should stop at this point. This should on
average speed up checking for emptiness by 2×.

Counting neighbours is used only by the unit test, and is retained for
that purpose.