Clamp output when outside type bounds
Previously the Op would recompute the calculation whenver the Mersenne
Twister provided a random value that would result in the input going
outside of the type bounds, in the hopes that another computation would
provide a result that was inside the type bounds. Not only can this be
slow, but also does not line up with what the user might expect.

Suppose that the input to the Op is an UnsignedByteType of value 254,
	with rangeMin=0 and rangeMax=4. Since we have uniform noise, we
	would expect that the probability that the output has an equal
	probability of being 254, 255, 256, 257, 258 (with a probability
	of 0.2 for each). But we know that this UnsignedByteType cannot
	hold a value greater than 255. If the Op redoes the calculation
	every time the output is greater than 255, then we know that the
	output will be 254 half of the time and 255 half of the time.
	Therefore no probability matches what might be expected.

Consider the same example, but instead of a recomputation we clamp the
output when it is outside the type bounds. We now have
P(output=254)=0.2, P(output=255)=0.8. In this case, all outputs within
the type range will have the expected probability of occurring, which
makes more sense than the scenario explained above. This is why the
change was made.