Separate Op into integer case and realType case
This allows us to handle integer types and realTypes separately. The
reason for this is as follows:

It seems clear that we would want to allow floating point randomness
when our type allows floating points

If we always base our calculations on floating points, this results in
edge effects when we add noise to integer types. This is due to
rounding: suppose we have a pseudorandomly generated double between 0
and 2. Most IntegerTypes simply round that double to a long, integer,
    etc in setReal. Thus the probability we add 0 to the input is 0.25,
    1 to the input is 0.5, and 2 to the input is 0.25. To get around
    this case we want to generate the random number differently when we
    have an integer type.

For these reasons we have two different Ops to perform the different
noise generation and have shared logic for setting the output.

N.B. We do not allow wrapping for floating point types. This is because
we operate within the following constraint: the number after the max of
the type should map to the minimum of the type. In the case of floating
points, what then should happen to a number equal to 0.5 + the maximum
of the type? In this paradigm, we would expect it to be the minimum of
the type - 0.5; this is impossible, however, since this would be outside
of the range of the type. For this reason, we always clamp.