Group Ops search results into sets
This change eliminates duplicate search results by grouping op
implementations by algorithm, folding all implementations of each
algorithm into one search result.

In ImageJ Ops, we can write implementations with overloading signatures;
we then let the Op matcher choose the best implementation for our
inputs. Thus the most consistent folding behavior would align with the
folding behavior of the matcher. In other words, groups should share:

* Op name
* Op functional type
* Simplified parameter types for *all* inputs and outputs (such as
  image or number)

Ideally, we would also be able to fold secondary parameters into a
single search result; however, this commit does not introduce that
functionality. I avoided this functionality knowing that SciJava Ops
does not differentiate between primary and secondary parameters.
Implementing secondary parameter folding would likely result in
inconsistent results between ImageJ Ops and ImageJ Ops2.

To make these conditions concrete, we introduce OpListing, a class
designed to encapsulate the information common to a set of
implementations. Each OpInfo should be able to provide an OpListing,
and Ops that are different implementations of the same algorithm should
provide the same OpListing.

We also introduce the OpListingModule, which can execute any known Ops
described by OpListing given suitable inputs, and the OpListingInfo that
can create the OpListingModule. These classes allow OpListings to be
runnable in ImageJ2 just like Ops and Modules, along with access to the
pre- and postprocessor chain of the Context.

Before this change, the OpSearcher would return one result for each
*implementation* available containing the searched text in its name.
With this change, results are grouped by OpListing. Simplified types are
provided to the OpSearcher, allowing the folding to occur in an
extensible way, with easy support for simplified type addition. For each
group, only one OpInfo's OpListing is chosen; this OpListing is
responsible for delegating to all Ops in the set.

Signed-off-by: Curtis Rueden <ctrueden@wisc.edu>